Update patch set 1
Patch Set 1:

INT-6 provided for tenant name to be specified if different than the default tenant for the credentials. This appears to be full automation of INT-6 and mitigates the need for GLOBAL_OPENSTACK_TENANT_NAME variable. The properties should be updated to remove references to GLOBAL_OPENSTACK_TENANT_NAME to avoid confusion.

Patch-set: 1