Update patch set 1
Patch Set 1:

Hi Vijay. I'm author of this tests. What do you mean that there is dedicated service for heltchecks within DCAE? You mean that all new components which are inside DCAE shouyld not have separate endpoint, so there is need to change code of this components? Or it's related to testing strategy and you mean that tests should be changed?

Patch-set: 1