Update patch set 3
Patch Set 3:

Regarding the csit filename prefix, I want to make it clear that they are different because the underlying functions are being rewritten. They are easily renamed afterwards.

However, I am not intending to rewrite the demo functions in this commit, because I am not testing the demo scenario. That would be yak-shaving from my point of view at this time.

The commit to add the new functions and the commits that rewrite demo functions are probably better done in separate steps.

The csit functions will manipulate certain objects in ways that exercise both negative (intended-to-fail) tests as well as positive (intended-to-pass) tests.

The AAI demo functions are for setup purposes, aren't they? The demo functions will probably want to wrap the specific scenario in easy-to-use keywords that are always intended to pass and continue to have their specific values and relationships set.

The actual demo test scenario itself is every other ONAP component performing their functions. That part, right now, is not of my concern.

I simply want to avoid unintended breakage of areas I don't want to touch.

Patch-set: 3
Reviewer: Gerrit User 4341 <4341@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0