Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Morgan - I have reviewed the proposition you have posted on onap-discuss and I am aligned with the strategy for Frankfurt

Daniel merged quickly, we could have waited for the PTL meeting..but we can revert if needed.
I also think it is a good compromise.
In this patch I did not exclude modeling - I need to contact the PTL. The project got a GO for frankfurt but their unique test has been very regularly failing in El Alto and in master - seems linked to env and race conditions.

Patch-set: 1