Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > what is the master robot lint job ? it seems to be failing due to tooling
> > 
> > 12:05:52   warnings.warn(warning, RequestsDependencyWarning)
> > 12:05:52 ERROR: Package 'zipp' requires a different Python: 3.5.2 not in '>=3.6'
> > 12:05:52 [testsuite-master-robot-lint] $ /bin/bash /tmp/jenkins8772115590667210864.sh
> 
> hmm I add a robotframeworklinter but theoretically, the target shall not be in the target env (just for local use)..so it shall not be triggered..I have a look

ok it is not in the env target (and not merged) but a trigger is already set in ci-management, which works well..it detects a change in a robot file, it is trying to launch the associated linter.

As for the moment there are too many Errors according to the linter, I will remove the trigger from ci-management.
it still could be launched locally once https://gerrit.onap.org/r/c/testsuite/+/103291 is merged but it will nto be triggered

Patch-set: 1