Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> So will aai oom project repo be moved back into oom as well ? the whole idea was to empower the projects to do their own helm charts as consistent updates so a project config change and matching helm charts could be done as one thing by the project - seems like moving the health checks into the project repo would be a better way to match health check changes with the project changes ?

It is like git submodules, it sounds like a good idea but it is not really at the end. 😊
when there is a change in AAI impacting a test, we need to sync 3 repos, which is not possible
moreover the skills in helm charts are limited, if you allow all the projects to do their charts, it become unmanageable for the OOM team (hardcoded passwords, templating, security, linting, divergence, lots of DBs/versions versus unification effort, docker optimization,..)

You are right, projects would probably feel more committed to the healthcheck tests if they had to host them in their repository. I think integration could suggest that for Guiling. Another option is to keep the robot part in testsuite but get new healthcheck from the projects (like policy and nbi)

Patch-set: 2