Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> So will aai oom project repo be moved back into oom as well ? the whole idea was to empower the projects to do their own helm charts as consistent updates so a project config change and matching helm charts could be done as one thing by the project - seems like moving the health checks into the project repo would be a better way to match health check changes with the project changes ?


That's my intention yes.

When you look on how a simple Dockerfile is done by some projects (double processes, "I install curl and nc cause I like to debug even in production", ...), I'm pretty sure that no project is able to handle correctly its own helm charts.

They still can propose changes in OOM but Krzysztof (O. not K. ;) ) and I will be able to take a look and verify that all the charts are (becoming) consistents.

We've got a long way before the project becomes cloud natives and their charts are super simple. Look at SO for example, all the projects are in the same git repo, meaning it's impossible to update one without building the others. This is not "cloud native".

And we have countless examples on ONAP these days.
When we'll have strong components, reliable deployments, it will be time to see how we can give more freedom for the project.

But I'm sure that today it will be a mistake (and AAI breaking once in a while -- although a lot less now they wait for us to say +1 -- is the perfect proof for me)

Patch-set: 2