Update patch set 2
Patch Set 2:

So good discussion.

I have seen oom helm charts break projects since the PTL is not always included in chart changes that cause their projects to break so probably want to tighten that process up a bit.

That being said - robot is broken - distribution and instantiation is blocked for Guilan.

When will a fix be in place so that we have a docker container that works for healthdist, isntantiateVWCL , instantiateVFWDT, instantiatevCPEInfra , etc ?

Patch-set: 2