Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > So good discussion.
> > 
> > I have seen oom helm charts break projects since the PTL is not always included in chart changes that cause their projects to break so probably want to tighten that process up a bit.
> > 
> > That being said - robot is broken - distribution and instantiation is blocked for Guilan.
> > 
> > When will a fix be in place so that we have a docker container that works for healthdist, isntantiateVWCL , instantiateVFWDT, instantiatevCPEInfra , etc ?
> 
> Sorry Frankfurt release not Guillan. Seems like a Sev 1 Test Blocker if we cant do any automated testing.

today healthdist is successuly run on OOM Master (pre-frankfurt) everyday and on every gating so we can do some automatic testing but we need a robot pod in good shape to use the other scripts especially the the 3 ones you are mentioning.

I would like to run isntantiateVWCL, instantiateVFWDT, instantiatevCPEInfra the same way (as pnf_registrate)
But I think these use cases are dependending on the installation (need networks and preconditions done through the onap_overide.yaml) and cannot be run as a simple robot tag. Do you confirm that demo init is needed prior to the 3 instantiation tests.

I think the patch I submitted should fix the 3rd party repo existing issue.
Kryzstof K. made recent tests and was able to run demo init.

I do not clearly see how far we are if we move to Daniel new docker (as far as I understand in his mail, it is already used in ATT, so we shall not be very far)
I hope I could give a try tomorrow.
We should decide quickly if it is achievable for Frankfurt..or not.

Regarding the problem of instantiation, I am not sure it is 100% due to the robot at the moment - we got also an issue when performing E2E tests (basic_vm) with onap-test framework (working fine in El Alto). I suspect some changes in A&AI/SDNC (as we had in SDC with the new onboarding procedure). we got a SDNC/AAI issue on module creation.

Patch-set: 2