Update patch set 1
Patch Set 1:

I tested the script, it works fine if you are on an existing remote branch (first try on a temp branch based on the patchset => fail).
It is a bash script to help building the docker anywhere and it works well.

As indicated I still prefer doing all the clone operations within the Docker file but I understand your approach (you did not touch to the existing stuff) you find a way to do what is done in jenkins (and should not be done in jenkins...)

I updated a patchset to do everything in the Docker file (https://gerrit.onap.org/r/c/testsuite/+/105168)
With this Dockerfile, you do not need to clone anything on your side, the Dockerfile must do it either through an installation of a python module or as a clone.
The ci-management, could be simplified (clone operations are not needed in jenkins anymore)

so let's wait weekly meeting discussion tomorrow. As the patch mentioned aboved is temporarily if we adopt Daniel's new version, but is it possible for Frankfurt?

Anyway your patch was useful to build the docker as it was defined, it could have been merged.
Sorry I did not see it earlier.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0