Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> isnt the whole point of init to do this setup?

init is doing a big init for all the use cases based on config set in the override.yaml file used by the robot pod during installation

for the use cases in CI, we need to keep the time under control and cannot perform a global init, which is long and setup lots of things we do not need.
And when like here no OpenStack connectivity is really required, it can speed up the test and focus on the purpose of the tests.
such evolution may be needed for the tests integrated in CI but should not break anything when dealing with the init part
let's give a try

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>