Update patch set 1
Patch Set 1: Code-Review-1

Could you be more explicit on the title..
explaining what you are fixing
- the use case
- the module (I assume it if for a CDS related test)
- the target release

this patch on testsuite is done as we are closed to the RC2..does it mean that we have to release a new version of testsuite for Frankfurt? is it needed for Frankfurt?

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1