Update patch set 3
Patch Set 3: Verified+1 Code-Review+2

thanks it is more explicit now

If we want this change integrated in frankfurt robot pod, we would need to self release a new version of robot.
it is not difficult but it takes some time...and we are closed to RC2.

Patch-set: 3
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2
Label: Verified=+1