Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Morgan indicates the error may be because of parallel robot's running to making the client ID based on hostname which is the container ID and should be unique

Yes I think it is linked to the // and the fact that the tests are sending same requests which is not planned by the components, we have a similar issue on SDC.
We can observe tha this error have not occured on the daily chains for several days now..so it is hard to reproduce.
You patch anyway may improve the test
Did you test it on a lab?

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>