Update patch set 1
Patch Set 1:

I don't think the problem was fixed yesterday with Nagendra's second review, although somehow the problems is masked in the report in error.

One problem is that URL is wrong, it is using a non-existing service:port URL that does not exist (pdp:8081).   There is no such service.

Note that there is no assertion test reported in the output.

Check out the section of the "Run Delete Policy Request" section of the policy section to see these issues:

https://logs.onap.org/onap-integration/daily/onap_daily_pod4_master/09-11-2020_07-09/xtesting-healthcheck/full/full/log.html#s1-t23

Patch-set: 1