Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> (3 comments)
> 
> ok, looks like just one evaluate issue left.

We have tried with all combinations for reading .zip file in a variable.. we have tried with Set Variable, Get Binary File, Set File and with encoding in the code as well. However its not working to substitute evaluate.. Can we make evaluate as it is and proceed ahead? In the next patch we can substitute evaluate if till the time we get substitution for it.

Patch-set: 5