Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5: Code-Review-1
> > > 
> > > (3 comments)
> > > 
> > > ok, looks like just one evaluate issue left.
> > 
> > We have tried with all combinations for reading .zip file in a variable.. we have tried with Set Variable, Get Binary File, Set File and with encoding in the code as well. However its not working to substitute evaluate.. Can we make evaluate as it is and proceed ahead? In the next patch we can substitute evaluate if till the time we get substitution for it.
> 
> this works in other areas. see https://gerrit.onap.org/r/gitweb?p=testsuite.git;a=blob;f=robot/resources/vnfsdk_interface.robot;h=c54b330e9b870b9b34a83bb074f5d64e1ac1b93d;hb=92b3b581276247ed828a33bc8e35d84b0bf31e5f


we tried again set variable, set file and get binary file as a substitute but it is working fine to read json file but not for reading .zip file.. can we verify this code with evaluate? As you provided the reference link, we have gone through it but if we pass file as an argument in the reference file and eventually we have to use evaluate in the test suite file to pass as an argument.

Patch-set: 5