Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > Are there any more comments or can we merge this change?
> > > 
> > > does it fix the regression observed on hv-ves on the recent daily chains?
> > 
> > It does fix only the issue with config map finding. To make the new hv-ves test pass, after we merge this, we will need also need to update hv-ves version in OOM: https://gerrit.onap.org/r/c/oom/+/123893
> 
> Let's merge it and check the gating of the OOM https://gerrit.onap.org/r/c/oom/+/123893 change - if test passes we will have a green light to update the OOM and master. On the other case we should revert these two changes due to regression on the tests, do you agree?

If after this change is merged, HV-VES tests still fail and there is no time to fix it, then yes, it probably should be reverted. The problem is that the new HV-VES uses configuration from config map, not Consul, so after changing it in the tests they are expected to fail with the old HV-VES container. If everything works fine https://gerrit.onap.org/r/c/oom/+/123893 should fix the tests.

Patch-set: 3