Added improvement in enhanced policy health check
due to timeout issue for the enhanced policy HC errors are seen that it sometimes takes long to undeploy and delete the policy. It results in incomplete cleanup and hinders the next run. So, added cleanup in the beginning of the test for any operation for cleanup left in the last run.


Issue-ID: POLICY-3887
Signed-off-by: Rommel Pawar <rommel.pawar@bell.ca>
Change-Id: I53443414ee71a79ac88c89fe2259f7ce8435780a